text.js created an functions moved
